bot/requirements.txt
python-telegram-bot[job-queue]>=20.7
gspread>=5.12.0
google-auth>=2.23.0
python-dotenv>=1.0.0
requests>=2.31.0
